% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lg2_mods.R
\name{lg2_mods}
\alias{lg2_mods}
\alias{disc_log}
\alias{disc_log_E}
\alias{#'}
\alias{@aliases}
\alias{disc_LV_comp}
\alias{disc_LV_E}
\title{Discrete-time logistic population growth, with options for competition, environmental covariate, and trait evolution}
\usage{
disc_log(r, N0, alpha)

disc_log_E(r, N0, alpha, E, x)

disc_LV_comp(r, N0, alpha)

disc_LV_E(r, N0, alpha, E, x)
}
\arguments{
\item{r}{the user-supplied population growth rate (when the species trait x is at the environmental optimum)}

\item{N0}{the initial population size}

\item{alpha}{the intraspecific interaction coefficient}

\item{E}{the initial value of the environmental covariate}

\item{x}{the initial value of the species trait}
}
\value{
xx
}
\description{
disc_log and disc_log_E are functions for logistic growth in a single species, with user-provided growth rate, initial population size, and strength of intraspecific competition. The function disc_log_E includes an environmental covariate, and growth rate now depends on this covariate as a function of the distance from the optimal trait value. XX disc_LV_comp XX disc_LV_E XX
}
\examples{
# Simulate initial species population growth
N1.0 <- rpois(1,10)
r1.0 <- 1.67
alpha.11 <- 0.00125
# Simulation of model for t time steps
t <- 30
N <- rep(NA, t)
N[1] <- N1.0
for (i in 2:t) {
  N[i] <- disc_log(r=r1.0, N0=N[i-1], alpha=alpha.11)
}

}
