% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lg3_mod.R
\name{lg3_mod}
\alias{lg3_mod}
\title{Discrete-time evolutionary rescuse + competition model for three species}
\usage{
lg3_mod(n, E, tx, parms)
}
\arguments{
\item{n}{A vector with the population size for all three species (N1, N2, N3) at time t}

\item{E}{A scalar with the local environmental optimum trait value E}

\item{tx}{A vector with the trait value for all three species (x1, x2, x3) at time t}

\item{parms}{A vector with the elements:
\describe{
\item{"a11"}{Intraspecific competition coefficient for species 1}
\item{"a12"}{Per-capita impact of species 2 on species 1}
\item{"a13"}{Per-capita impact of species 3 on species 1}
\item{"a21"}{Per-capita impact of species 1 on species 2}
\item{"a22"}{Intraspecific competition coefficient for species 2}
\item{"a23"}{Per-capita impact of species 3 on species 2}
\item{"a31"}{Per-capita impact of species 1 on species 3}
\item{"a32"}{Per-capita impact of species 2 on species 3}
\item{"a33"}{Intraspecific competition coefficient for species 1}
\item{"wmax"}{The maximum growth rate of a species}
\item{"w"}{The strength of selection, or the width of the trait-fitness Gaussian distribution}
\item{"P"}{The width of the phenotypic distribution}
\item{"h21"}{Heritability of trait x in species 1}
\item{"h22"}{Heritability of trait x in species 2}
\item{"h23"}{Heritability of trait x in species 3}
}}
}
\value{
The values of population size, trait values, and population growth rate for species 1, 2, and 3.
}
\description{
Equations for Leslie-gower population growth (discrete-time Lotka-Volterra) models for three species. Population growth is modelled as population mean fitness as per Gomulkiewicz & Holt 1995. These formulas calculate the population's maximum growth rate (with a trait value at the local environmental optimum), the distance from the optimal trait value and evolutionary inertia of trait given phenotypic variation, and the change in population size N from one time step to the next given intra- and interspecific interaction strength.
}
\details{
The model is a discrete time, stochastic, quantitative genetic model of evolutionary rescue (Gomulkiewicz & Holt 1995), with some modifications. Populations experience logistic growth: \eqn{N_t+1 = N_t + \lambda N_t ((1-N_t)/K)} and the population growth rate \eqn{\lambda} is the absolute mean fitness of a population in generation \emph{t}. Mean fitness is determined by a quantitative trait \emph{x} as follows:
\eqn{\lambda = \hat{W}e^{w+(1-h^2)P / (P+w)(E-x_t)]^2}{2(P+w)}}

where \eqn{\hat{W} = W_max sqrt(\frac{w}{P + w})}, \emph{W_max} is the absolute fitness when the phenotype \emph{x} equals the local environmental optimum trait value \emph{E}, \emph{w} is the width of the Gaussian fitness function (which determines the strength of selection), \emph{P} is the width of the distribution of the phenotype \emph{x}, $h^2$ is the heritability of the phenotype \emph{x}, and \emph{x_t} is the population's mean phenotype at time \emph{t}.
}
\examples{
sim_num <- 10
m1 <- rand_sim(n=sim_num,x0=c(0.4,0.5,0.6))
## Visualize the results
matplot(1:300, m1$pop_record[1,,7:9], type = "l", xlab="time",ylab="N")
}
