% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lk_evol.R
\name{lk_evol}
\alias{lk_evol}
\title{Function with ODE system for host-pathogen eco-evolutionary model}
\usage{
lk_evol(t, x, params)
}
\arguments{
\item{t}{A vector of time steps for which the differential equations will be evaluated across}

\item{x}{A vector of starting values for S, I, and alpha}

\item{params}{A vector of values for the model parameters B, μ, V, γ, and c}
}
\value{
The values of the derivatives in the ODE system at time t.
}
\description{
Function with ODE system for host-pathogen eco-evolutionary model
}
\examples{
parms <- c(B = (10^7), mu = 1, V = 0, gam = 2, c = 3*10^-7)
times <- seq(from=0,to=40,by=1)
xstart <- c(S=10^7,I=1,alpha=4)
out <- hostpath_evol(times,xstart,parms)

}
