% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lk_mods.R
\name{lk_mods}
\alias{lk_mods}
\alias{lk_ecoXevo}
\alias{lk_eco_evo}
\title{Functions with ODE system for host-pathogen eco-evolutionary model}
\usage{
lk_ecoXevo(t, x, params)

lk_eco_evo(t, x, params)
}
\arguments{
\item{t}{A vector of time steps for which the differential equations will be evaluated across}

\item{x}{A vector of starting values for S, I, and alpha}

\item{params}{A vector of values for the model parameters B, μ, V, γ, and c}
}
\value{
The values of the derivatives in the ODE system at time t.
}
\description{
lk_ecoXevo and lk_eco_evo are functions with ordinary differential equations for an eco-evolutionary host pathogen model. They differ in whether virulence evolution evolves and impacts host population size S - in lk_ecoXevo these are coupled and in lk_eco_evo virulence evolution is not dependent on S.
}
\examples{
parms <- c(B = (10^7), mu = 1, V = 0, gam = 2, c = 3*10^-7)
times <- seq(from=0,to=40,by=1)
xstart <- c(S=10^7,I=1,alpha=4)
out <- hostpath_evol(times,xstart,parms,method="ecoXevo")

}
