% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LV_evol.R
\name{LV_evol}
\alias{LV_evol}
\title{Simulate three-species discrete-time growth with competition and evolving growth rate
Simulate competition, growth, and trait evolution in an eco-evolutionary version of a discrte-time Leslie-Gower model. The function calls the formulas in "lg3_mod" over the time intervals for which population growth is calculated.}
\usage{
LV_evol(N0, x0, E, time_points, parms)
}
\arguments{
\item{N0}{A vector with the initial population size for all three species ($N_{1,0}$, $N_{2,0}$, $N_{3,0}$)}

\item{x0}{A vector with the initial trait value \emph{x} for all three species ($x_{1,0}$, $x_{2,0}$, $x_{3,0}$)}

\item{E}{A scalar with the local environmental optimum trait value E}

\item{time_points}{A numeric value of the number of time steps for the simulation}

\item{parms}{A vector with the elements:
\itemize{
\item{"a11"}{Intraspecific competition coefficient for species 1}
\item{"a12"}{Per-capita impact of species 2 on species 1}
\item{"a13"}{Per-capita impact of species 3 on species 1}
\item{"a21"}{Per-capita impact of species 1 on species 2}
\item{"a22"}{Intraspecific competition coefficient for species 2}
\item{"a23"}{Per-capita impact of species 3 on species 2}
\item{"a31"}{Per-capita impact of species 1 on species 3}
\item{"a32"}{Per-capita impact of species 2 on species 3}
\item{"a33"}{Intraspecific competition coefficient for species 1}
\item{"wmax"}{The maximum growth rate of a species}
\item{"w"}{The strength of selection, or the width of the trait-fitness Gaussian distribution}
\item{"P"}{The width of the phenotypic distribution}
\item{"h21"}{Heritability of trait x in species 1}
\item{"h22"}{Heritability of trait x in species 2}
\item{"h23"}{Heritability of trait x in species 3}
}}
}
\value{
A dataframe with the elements:
\itemize{
\item{"x1,x2,x3"}{The population size of all three species at time t}
\item{"tx1,tx2,tx3"}{The trait value of all three species at time t}
\item{"y1,y2,y3"}{The population size of all three species at time t+1}
\item{"yt1,yt2,yt3"}{The trait value of all three species at time t+1}
\item{"E"}{The local environmental optimum trait value at time t+1}
\item{"l1,l2,l3"}{The per-capita  growth rate of all three species from time t to t+1}
}
\description{
Simulate three-species discrete-time growth with competition and evolving growth rate
Simulate competition, growth, and trait evolution in an eco-evolutionary version of a discrte-time Leslie-Gower model. The function calls the formulas in "lg3_mod" over the time intervals for which population growth is calculated.
}
\examples{
sim_num <- 10
m1 <- rand_sim(n=sim_num,x0=c(0.4,0.5,0.6))
## Visualize the results
matplot(1:300, m1$pop_record[1,,7:9], type = "l", xlab="time",ylab="N")
}
