% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gen3sis.R
\name{plot_gen3sis}
\alias{plot_gen3sis}
\title{Plot gen3sis results, compare across multiple simulation runs}
\usage{
plot_gen3sis(
  output,
  summary_title = NULL,
  summary_legend = NULL,
  max_gam = 500,
  max_alpha = 100
)
}
\arguments{
\item{output}{a sgen3sis output object resulting from a gen3sis simulation (i.e. run_simulation; see gen3sis::run_simulation)}

\item{summary_title}{summary plot title as character. If NULL, title is computed from input name (see gen3sis::run_simulation)}

\item{summary_legend}{either a staring with \verb{\\_n} for new lines or NULL. If NULL, provides default summary and simulation information (see gen3sis::run_simulation)}

\item{max_gam}{A maximum gamma diversity value across multiple simulation runs. Sets the right-y-axis for the gamma diversity plot.}

\item{max_alpha}{A maximum alpha diversity value across multiple simulation runs. Sets the color scale for the alpha diversity map plot.}
}
\value{
no values returned, only for plotting
}
\description{
Plot gen3sis results, compare across multiple simulation runs
}
\examples{
# load existing summary example
datapath <- system.file(file.path("extdata", "WorldCenter"), package = "gen3sis")
output <- readRDS(file.path(datapath, "output/config_worldcenter/sgen3sis.rds"))
# plot output summary
max_gam <- max(output$summary$phylo_summary[,"alive"])
max_alpha <- max(output$summary$`richness-final`[,3],na.rm=TRUE)
plot_gen3sis(output,max_gam=max_gam, max_alpha=max_alpha)
}
