% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand_sim.R
\name{rand_sim}
\alias{rand_sim}
\title{Run a simulation with parameter values randomly chosen from prior distributions (currently for 3-species Leslie-Gower model)}
\usage{
rand_sim(
  N0 = c(10, 10, 10),
  x0 = NULL,
  E = 0.1,
  time_points = 300,
  n,
  evol = FALSE
)
}
\arguments{
\item{N0}{A vector with the initial population size for all three species ($N_{1,0}$, $N_{2,0}$, $N_{3,0}$)}

\item{x0}{A vector with the initial trait value \emph{x} for all three species ($x_{1,0}$, $x_{2,0}$, $x_{3,0}$)}

\item{E}{A scalar with the local environmental optimum trait value E}

\item{time_points}{A numeric value of the number of time steps for the simulation}

\item{n}{A numeric value of the number of simulations to run}

\item{evol}{A boolean where FALSE sets the heritability for all species equal to 0 and TRUE draws a heritability value for all species from a uniform distribution between 0-1.}
}
\value{
A list with the elements:
\itemize{
\item{"pop_record"}{A three-dimensional array that returns the following for all simulation iterations (n) at all time points:
\describe{
\item{"x1,x2,x3"}{The population size of all three species at time t}
\item{"tx1,tx2,tx3"}{The trait value of all three species at time t}
\item{"y1,y2,y3"}{The population size of all three species at time t+1}
\item{"yt1,yt2,yt3"}{The trait value of all three species at time t+1}
\item{"E"}{The local environmental optimum trait value at time t+1}
\item{"l1,l2,l3"}{The per-capita  growth rate of all three species from time t to t+1}
}}
\item{"parm_record"}{An n-by-18 matrix with the values for all model parameters used for each of the n simulation iterations.}
}
\description{
Run a simulation with parameter values randomly chosen from prior distributions (currently for 3-species Leslie-Gower model)
}
\examples{
sim_num <- 10
m1 <- rand_sim(n=sim_num,x0=c(0.4,0.5,0.6))
## Visualize the results
matplot(1:300, m1$pop_record[1,,7:9], type = "l", xlab="time",ylab="N")
}
