% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rf_lv}
\alias{rf_lv}
\title{Random forest feature selection, parameter tuning, and model classification for simulations of 3-species Lotka-Volterra model}
\format{
A list with the following elements:
\describe{
\item{bor}{A list with the elements resulting from feature selection algorithm Boruta::Boruta}
\item{tune}{A list with the elements resulting from parameter tuning randomForest:tuneRF}
\item{rf}{A list with the elements resulting from parameter tuning randomForest:randomForest}
}
}
\source{
See ecoevo-ABC section '2 Simulation'
}
\usage{
rf_lv
}
\description{
A dataset containing the results of a boruta feature selection from simulated data of 3 species population size and trait values over time (Boruta::Boruta), parameter tuning (randomForest:tuneRF), and model classification (randomForest:randomForest). Simulated data comes from a 3-species Lotka-Volterra model with a quantitative genetic model for potential evolution of population growth rate (see vignette ecoevo-ABC). Hypotheses for model classification of simulated observed data are No Evolution(h2 = 0) and Evolution (h2 > 0).
}
\keyword{datasets}
