% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_up.R
\name{round_up}
\alias{round_up}
\title{round a number up to the nearest power of 10}
\usage{
round_up(x)
}
\arguments{
\item{x}{the number to round}
}
\value{
the rounded value
}
\description{
round a number up to the nearest power of 10
}
\examples{
round_up(9)
round_up(1020)
}
